+++ plexapi/src/main/java/com/liskovsoft/plexapi/service/PlexSignInServiceImpl.java
@@ -63,6 +63,7 @@ public class PlexSignInServiceImpl implements PlexSignInService {
     @Override
     public void signOut() {
         prefs().clearAuthToken();
+        prefs().clearSelectedServer();
     }
 
     @Override